Audio: Move FIR core to math library
This patch enables FIR filter core usage independently from FIR
equalizer component. The inline of FIR core is removed to reduce
the code size when FIR is used from several components. Each
component also typically used inline FIR version for each supported
PCM format that increased further the size.

Most of the changes are due to rename and directory move of some FIR
data structures and macros. The code functionality is not changed.
