+++ src/audio/eq_fir/eq_fir_hifi3.c
@@ -4,87 +4,25 @@
 //
 // Author: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>
 
-#include <sof/audio/eq_fir/fir_config.h>
+#include <sof/math/fir_config.h>
 
 #if FIR_HIFI3
 
-#include <sof/audio/buffer.h>
-#include <sof/audio/eq_fir/fir_hifi3.h>
-#include <user/eq.h>
+#include <sof/audio/eq_fir/eq_fir.h>
+#include <sof/math/fir_hifi3.h>
+#include <user/fir.h>
 #include <xtensa/config/defs.h>
 #include <xtensa/tie/xt_hifi3.h>
 #include <errno.h>
 #include <stddef.h>
 #include <stdint.h>
 
-/*
- * EQ FIR algorithm code
- */
-
-void fir_reset(struct fir_state_32x16 *fir)
-{
-	fir->taps = 0;
-	fir->length = 0;
-	fir->out_shift = 0;
-	fir->coef = NULL;
-	/* There may need to know the beginning of dynamic allocation after
-	 * reset so omitting setting also fir->delay to NULL.
-	 */
-}
-
-int fir_delay_size(struct sof_eq_fir_coef_data *config)
-{
-	/* Check FIR tap count for implementation specific constraints */
-	if (config->length > SOF_EQ_FIR_MAX_LENGTH || config->length < 4)
-		return -EINVAL;
-
-	/* The optimization requires the tap count to be multiple of four */
-	if (config->length & 0x3)
-		return -EINVAL;
-
-	/* The dual sample version needs one more delay entry. To preserve
-	 * align for 64 bits need to add two.
-	 */
-	return (config->length + 2) * sizeof(int32_t);
-}
-
-int fir_init_coef(struct fir_state_32x16 *fir,
-		  struct sof_eq_fir_coef_data *config)
-{
-	/* The length is taps plus two since the filter computes two
-	 * samples per call. Length plus one would be minimum but the add
-	 * must be even. The even length is needed for 64 bit loads from delay
-	 * lines with 32 bit samples.
-	 */
-	fir->taps = (int)config->length;
-	fir->length = fir->taps + 2;
-	fir->out_shift = (int)config->out_shift;
-	fir->coef = (ae_f16x4 *)&config->coef[0];
-	return 0;
-}
-
-void fir_init_delay(struct fir_state_32x16 *fir, int32_t **data)
-{
-	fir->delay = (ae_int32 *) *data;
-	fir->delay_end = fir->delay + fir->length;
-	fir->rwp = (ae_int32 *)(fir->delay + fir->length - 1);
-	*data += fir->length; /* Point to next delay line start */
-}
-
-void fir_get_lrshifts(struct fir_state_32x16 *fir, int *lshift,
-		      int *rshift)
-{
-	*lshift = (fir->out_shift < 0) ? -fir->out_shift : 0;
-	*rshift = (fir->out_shift > 0) ? fir->out_shift : 0;
-}
-
 #if CONFIG_FORMAT_S32LE
 /* For even frame lengths use FIR filter that processes two sequential
  * sample per call.
  */
-void eq_fir_2x_s32_hifi3(struct fir_state_32x16 fir[],
-			 const struct audio_stream *source,
-			 struct audio_stream *sink, int frames, int nch)
+void eq_fir_2x_s32(struct fir_state_32x16 fir[], const struct audio_stream *source,
+		   struct audio_stream *sink, int frames, int nch)
 {
 	struct fir_state_32x16 *f;
 	ae_int32x2 d0 = 0;
@@ -142,9 +80,8 @@ void eq_fir_2x_s32_hifi3(struct fir_state_32x16 fir[],
 }
 
 /* FIR for any number of frames */
-void eq_fir_s32_hifi3(struct fir_state_32x16 fir[],
-		      const struct audio_stream *source,
-		      struct audio_stream *sink, int frames, int nch)
+void eq_fir_s32(struct fir_state_32x16 fir[], const struct audio_stream *source,
+		struct audio_stream *sink, int frames, int nch)
 {
 	struct fir_state_32x16 *f;
 	ae_int32x2 in = 0;
@@ -200,9 +137,8 @@ void eq_fir_s32_hifi3(struct fir_state_32x16 fir[],
 #endif /* CONFIG_FORMAT_S32LE */
 
 #if CONFIG_FORMAT_S24LE
-void eq_fir_2x_s24_hifi3(struct fir_state_32x16 fir[],
-			 const struct audio_stream *source,
-			 struct audio_stream *sink, int frames, int nch)
+void eq_fir_2x_s24(struct fir_state_32x16 fir[], const struct audio_stream *source,
+		   struct audio_stream *sink, int frames, int nch)
 {
 	struct fir_state_32x16 *f;
 	ae_int32x2 d0 = 0;
@@ -265,9 +201,8 @@ void eq_fir_2x_s24_hifi3(struct fir_state_32x16 fir[],
 	}
 }
 
-void eq_fir_s24_hifi3(struct fir_state_32x16 fir[],
-		      const struct audio_stream *source,
-		      struct audio_stream *sink, int frames, int nch)
+void eq_fir_s24(struct fir_state_32x16 fir[], const struct audio_stream *source,
+		struct audio_stream *sink, int frames, int nch)
 {
 	struct fir_state_32x16 *f;
 	ae_int32 in;
@@ -328,9 +263,8 @@ void eq_fir_s24_hifi3(struct fir_state_32x16 fir[],
 #endif /* CONFIG_FORMAT_S24LE */
 
 #if CONFIG_FORMAT_S16LE
-void eq_fir_2x_s16_hifi3(struct fir_state_32x16 fir[],
-			 const struct audio_stream *source,
-			 struct audio_stream *sink, int frames, int nch)
+void eq_fir_2x_s16(struct fir_state_32x16 fir[], const struct audio_stream *source,
+		   struct audio_stream *sink, int frames, int nch)
 {
 	struct fir_state_32x16 *f;
 	ae_int16x4 d0 = AE_ZERO16();
@@ -397,9 +331,8 @@ void eq_fir_2x_s16_hifi3(struct fir_state_32x16 fir[],
 	}
 }
 
-void eq_fir_s16_hifi3(struct fir_state_32x16 fir[],
-		      const struct audio_stream *source,
-		      struct audio_stream *sink, int frames, int nch)
+void eq_fir_s16(struct fir_state_32x16 fir[], const struct audio_stream *source,
+		struct audio_stream *sink, int frames, int nch)
 {
 	struct fir_state_32x16 *f;
 	ae_f16x4 d = AE_ZERO16();
@@ -456,4 +389,5 @@ void eq_fir_s16_hifi3(struct fir_state_32x16 fir[],
 	}
 }
 #endif /* CONFIG_FORMAT_S16LE */
-#endif
+
+#endif /* FIR_HIFI3 */
