+++ src/math/CMakeLists.txt
@@ -1,9 +1,11 @@
 # SPDX-License-Identifier: BSD-3-Clause
 
-add_local_sources(sof numbers.c trig.c decibels.c iir_df2t_generic.c iir_df2t_hifi3.c)
-
 if(BUILD_LIBRARY)
 	return()
 endif()
 
-add_local_sources(sof trig.c decibels.c iir_df2t_generic.c iir_df2t_hifi3.c)
+add_local_sources(sof numbers.c trig.c decibels.c iir_df2t_generic.c iir_df2t_hifi3.c)
+
+if(CONFIG_MATH_FIR)
+        add_local_sources(sof fir_generic.c fir_hifi2ep.c fir_hifi3.c)
+endif()
