+++ src/math/fir_hifi3.c
@@ -1,99 +1,90 @@
-/* SPDX-License-Identifier: BSD-3-Clause
- *
- * Copyright(c) 2017 Intel Corporation. All rights reserved.
- *
- * Author: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>
- */
+// SPDX-License-Identifier: BSD-3-Clause
+//
+// Copyright(c) 2017 Intel Corporation. All rights reserved.
+//
+// Author: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>
 
-#ifndef __SOF_AUDIO_EQ_FIR_FIR_HIFI3_H__
-#define __SOF_AUDIO_EQ_FIR_FIR_HIFI3_H__
-
-#include <sof/audio/eq_fir/fir_config.h>
+#include <sof/math/fir_config.h>
 
 #if FIR_HIFI3
 
 #include <sof/audio/buffer.h>
+#include <sof/math/fir_hifi3.h>
+#include <user/fir.h>
 #include <xtensa/config/defs.h>
 #include <xtensa/tie/xt_hifi3.h>
+#include <errno.h>
+#include <stddef.h>
+#include <stdint.h>
 
-struct sof_eq_fir_coef_data;
+/*
+ * EQ FIR algorithm code
+ */
 
-struct fir_state_32x16 {
-	ae_int32 *rwp; /* Circular read and write pointer */
-	ae_int32 *delay; /* Pointer to FIR delay line */
-	ae_int32 *delay_end; /* Pointer to FIR delay line end */
-	ae_f16x4 *coef; /* Pointer to FIR coefficients */
-	int taps; /* Number of FIR taps */
-	int length; /* Number of FIR taps plus input length (even) */
-	int in_shift; /* Amount of right shifts at input */
-	int out_shift; /* Amount of right shifts at output */
-};
-
-void fir_reset(struct fir_state_32x16 *fir);
-
-int fir_delay_size(struct sof_eq_fir_coef_data *config);
-
-int fir_init_coef(struct fir_state_32x16 *fir,
-		  struct sof_eq_fir_coef_data *config);
-
-void fir_init_delay(struct fir_state_32x16 *fir, int32_t **data);
-
-#if CONFIG_FORMAT_S16LE
-void eq_fir_s16_hifi3(struct fir_state_32x16 *fir,
-		      const struct audio_stream *source,
-		      struct audio_stream *sink, int frames, int nch);
-
-void eq_fir_2x_s16_hifi3(struct fir_state_32x16 *fir,
-			 const struct audio_stream *source,
-			 struct audio_stream *sink, int frames, int nch);
-#endif /* CONFIG_FORMAT_S16LE */
-
-#if CONFIG_FORMAT_S24LE
-void eq_fir_s24_hifi3(struct fir_state_32x16 *fir,
-		      const struct audio_stream *source,
-		      struct audio_stream *sink, int frames, int nch);
-
-void eq_fir_2x_s24_hifi3(struct fir_state_32x16 *fir,
-			 const struct audio_stream *source,
-			 struct audio_stream *sink, int frames, int nch);
-#endif /* CONFIG_FORMAT_S24LE */
-
-#if CONFIG_FORMAT_S32LE
-void eq_fir_s32_hifi3(struct fir_state_32x16 *fir,
-		      const struct audio_stream *source,
-		      struct audio_stream *sink, int frames, int nch);
-
-void eq_fir_2x_s32_hifi3(struct fir_state_32x16 *fir,
-			 const struct audio_stream *source,
-			 struct audio_stream *sink, int frames, int nch);
-#endif /* CONFIG_FORMAT_S32LE */
-
-/* Setup circular buffer for FIR input data delay */
-static inline void fir_core_setup_circular(struct fir_state_32x16 *fir)
+void fir_reset(struct fir_state_32x16 *fir)
 {
-	AE_SETCBEGIN0(fir->delay);
-	AE_SETCEND0(fir->delay_end);
+	fir->taps = 0;
+	fir->length = 0;
+	fir->out_shift = 0;
+	fir->coef = NULL;
+	/* There may need to know the beginning of dynamic allocation after
+	 * reset so omitting setting also fir->delay to NULL.
+	 */
 }
 
-/* Setup circular for component buffer */
-static inline void fir_comp_setup_circular(const struct audio_stream *buffer)
+int fir_delay_size(struct sof_fir_coef_data *config)
 {
-	AE_SETCBEGIN0(buffer->addr);
-	AE_SETCEND0(buffer->end_addr);
+	/* Check FIR tap count for implementation specific constraints */
+	if (config->length > SOF_FIR_MAX_LENGTH || config->length < 4)
+		return -EINVAL;
+
+	/* The optimization requires the tap count to be multiple of four */
+	if (config->length & 0x3)
+		return -EINVAL;
+
+	/* The dual sample version needs one more delay entry. To preserve
+	 * align for 64 bits need to add two.
+	 */
+	return (config->length + 2) * sizeof(int32_t);
+}
+
+int fir_init_coef(struct fir_state_32x16 *fir,
+		  struct sof_fir_coef_data *config)
+{
+	/* The length is taps plus two since the filter computes two
+	 * samples per call. Length plus one would be minimum but the add
+	 * must be even. The even length is needed for 64 bit loads from delay
+	 * lines with 32 bit samples.
+	 */
+	fir->taps = (int)config->length;
+	fir->length = fir->taps + 2;
+	fir->out_shift = (int)config->out_shift;
+	fir->coef = (ae_f16x4 *)&config->coef[0];
+	return 0;
+}
+
+void fir_init_delay(struct fir_state_32x16 *fir, int32_t **data)
+{
+	fir->delay = (ae_int32 *)*data;
+	fir->delay_end = fir->delay + fir->length;
+	fir->rwp = (ae_int32 *)(fir->delay + fir->length - 1);
+	*data += fir->length; /* Point to next delay line start */
 }
 
 void fir_get_lrshifts(struct fir_state_32x16 *fir, int *lshift,
-		      int *rshift);
-
-/* The next functions are inlined to optmize execution speed */
+		      int *rshift)
+{
+	*lshift = (fir->out_shift < 0) ? -fir->out_shift : 0;
+	*rshift = (fir->out_shift > 0) ? fir->out_shift : 0;
+}
 
 /* HiFi EP has the follow number of reqisters that should not be exceeded
  * 4x 56 bit registers in register file Q
  * 8x 48 bit registers in register file P
  */
 
-static inline void fir_32x16_hifi3(struct fir_state_32x16 *fir, ae_int32 x,
-				   ae_int32 *y, int shift)
+void fir_32x16_hifi3(struct fir_state_32x16 *fir, ae_int32 x, ae_int32 *y,
+		     int shift)
 {
 	/* This function uses
 	 * 1x 56 bit registers Q,
@@ -170,9 +161,8 @@ static inline void fir_32x16_hifi3(struct fir_state_32x16 *fir, ae_int32 x,
  * 8x 48 bit registers in register file P
  */
 
-static inline void fir_32x16_2x_hifi3(struct fir_state_32x16 *fir, ae_int32 x0,
-				      ae_int32 x1, ae_int32 *y0, ae_int32 *y1,
-				      int shift)
+void fir_32x16_2x_hifi3(struct fir_state_32x16 *fir, ae_int32 x0, ae_int32 x1,
+			ae_int32 *y0, ae_int32 *y1, int shift)
 {
 	/* This function uses
 	 * 2x 56 bit registers Q,
@@ -256,4 +246,3 @@ static inline void fir_32x16_2x_hifi3(struct fir_state_32x16 *fir, ae_int32 x0,
 }
 
 #endif
-#endif /* __SOF_AUDIO_EQ_FIR_FIR_HIFI3_H__ */
